Use SPM instead of swiftc to build Swift templates
The process is now the following:
1. create the package folder hierarchy
2. synchronize the SourceryRuntime code into the package's `Sources` using the
embedded code in the Sourcery binary and `FolderSynchronizer`
3. write the manifest file `Package.swift`
4. delete `Sources/SwiftTemplate` folder if it exists to remove any file from
the previously built template
5. write the `main.swift` file in `Sources/SwiftTemplate`
6. add any included files in `Sources/SwiftTemplate`
7. build the template's binary using SPM

The hierarchy before building:
```
.
├── Package.swift
└── Sources
    ├── SourceryRuntime
    │   ├── ... (SourceryRuntime files)
    └── SwiftTemplate
        ├── main.swift
        └── ... (included files if any)
```

Things to note:
- this does not use the SourceryRuntime.framework anymore so this removes the
Swift version conflicts when using a different version of Xcode than the one
used to build Sourcery.
- the process do not remove the temporary folder anymore so that the
SourceryRuntime code is not rebuilt for every template. In the future, we'll
have to keep an eye on that because there may be some remaining files from one
run to the next. Right now this should not be a problem because all template
related files are removed at the end of `build()`.